Fix the endian read/writes in memorymappedFileByteBuffer

+++ test/core/Util/TestStringHelper.cs
@@ -27,10 +27,10 @@ namespace Lucene.Net.Util
         public virtual void TestMurmurHash3()
         {
             // Hashes computed using murmur3_32 from https://code.google.com/p/pyfasthash
-            Assert.AreEqual(0xf6a5c420, StringHelper.Murmurhash3_x86_32(new BytesRef("foo"), 0));
-            Assert.AreEqual(0xcd018ef6, StringHelper.Murmurhash3_x86_32(new BytesRef("foo"), 16));
-            Assert.AreEqual(0x111e7435, StringHelper.Murmurhash3_x86_32(new BytesRef("You want weapons? We're in a library! Books! The best weapons in the world!"), 0));
-            Assert.AreEqual(0x2c628cd0, StringHelper.Murmurhash3_x86_32(new BytesRef("You want weapons? We're in a library! Books! The best weapons in the world!"), 3476));
+            Assert.AreEqual(0xf6a5c420, (uint)StringHelper.Murmurhash3_x86_32(new BytesRef("foo"), 0));
+            Assert.AreEqual(0xcd018ef6, (uint)StringHelper.Murmurhash3_x86_32(new BytesRef("foo"), 16));
+            Assert.AreEqual(0x111e7435, (uint)StringHelper.Murmurhash3_x86_32(new BytesRef("You want weapons? We're in a library! Books! The best weapons in the world!"), 0));
+            Assert.AreEqual(0x2c628cd0, (uint)StringHelper.Murmurhash3_x86_32(new BytesRef("You want weapons? We're in a library! Books! The best weapons in the world!"), 3476));
         }
     }
